+++ pyccel/ast/internals.py
@@ -77,6 +77,17 @@ class PyccelFunction(TypedAstNode):
         """
         return ()
 
+    @property
+    def is_indexable(self):
+        """
+        Indicate whether the expression can be indexed.
+
+        Indicate whether the expression can be indexed to get an element without
+        calculating the entire result. E.g `cos(x)[i]` is equivalent to `cos(x[i])`
+        but `func_call(x)[i]` is not equivalent to `func_call(x[i])`.
+        """
+        return self.is_elemental
+
 class PyccelArraySize(PyccelFunction):
     """
     Gets the total number of elements in an array.