+++ pyccel/codegen/printing/ccode.py
@@ -18,7 +18,7 @@ from pyccel.ast.bind_c    import BindCPointer
 from pyccel.ast.builtins  import PythonRange, PythonComplex, PythonMin, PythonMax
 from pyccel.ast.builtins  import PythonPrint, PythonType, VariableIterator
 
-from pyccel.ast.builtins  import PythonList, PythonTuple, PythonSet, PythonDict, PythonLen
+from pyccel.ast.builtins  import PythonList, PythonTuple, PythonSet, PythonDict, PythonLen, PythonConjugate
 
 from pyccel.ast.builtin_methods.dict_methods  import DictItems, DictKeys, DictValues, DictPopitem
 
@@ -54,7 +54,7 @@ from pyccel.ast.mathext  import math_constants
 from pyccel.ast.numpyext import NumpyFull, NumpyArray, NumpySum, DtypePrecisionToCastFunction
 from pyccel.ast.numpyext import NumpyReal, NumpyImag, NumpyFloat
 from pyccel.ast.numpyext import NumpyAmin, NumpyAmax, NumpyAbs
-from pyccel.ast.numpyext import NumpyReduction
+from pyccel.ast.numpyext import NumpyReduction, NumpyMatmul
 from pyccel.ast.numpyext import get_shape_of_multi_level_container
 
 from pyccel.ast.numpytypes import NumpyFloat32Type, NumpyFloat64Type, NumpyFloat128Type
@@ -78,7 +78,8 @@ from pyccel.codegen.printing.codeprinter import CodePrinter
 
 from pyccel.errors.errors   import Errors
 from pyccel.errors.messages import (PYCCEL_RESTRICTION_TODO, INCOMPATIBLE_TYPEVAR_TO_FUNC,
-                                    PYCCEL_RESTRICTION_IS_ISNOT, PYCCEL_INTERNAL_ERROR)
+                                    PYCCEL_RESTRICTION_IS_ISNOT, PYCCEL_INTERNAL_ERROR,
+                                    ALLOCATABLE_IN_EXPRESSION)
 
 
 errors = Errors()
@@ -770,6 +771,80 @@ class CCodePrinter(CodePrinter):
             errors.report(f"The declaration of type {element_type} is not yet implemented for containers.",
                     symbol=expr, severity='error')
         return decl_line
+
+    def _inline_vecdot(self, expr, x1, x2, axis, *, with_conjugate):
+        """
+        Get the code describing a vector dot product.
+
+        Get the code describing a vector dot product with or without a possible
+        conjugate calculation for complexes. This is used for the implementation
+        of np.vecdot and np.matmul.
+
+        Parameters
+        ----------
+        expr : NumpyVecdot | NumpyMatmul
+            The node that is being printed.
+        x1 : TypedAstNode
+            The first vector.
+        x2 : TypedAstNode
+            The second vector.
+        axis : int
+            The axis along which the dot product is performed.
+        with_conjugate : bool
+            Indicates if the conjugate of the first argument should be used.
+
+        Returns
+        -------
+        str
+            The code that describes the vecdot operation.
+        """
+        assign_node = expr.get_direct_user_nodes(lambda p: isinstance(p, Assign))
+        if assign_node:
+            lhs_var = assign_node[0].lhs
+        else:
+            # Rank should always be 0 thanks to loop unrolling
+            assert expr.rank == 0
+            lhs_var = self.scope.get_temporary_variable(expr.class_type)
+        in_arg_vars = (x1, x2)
+        arg_vars = []
+        prefix = ''
+        for a in in_arg_vars:
+            if not isinstance(a, Variable):
+                # This handles slice arguments
+                assert a.rank
+                tmp = self.scope.get_temporary_variable(a.class_type, shape = a.shape,
+                        memory_handling='alias')
+                prefix += self._print(AliasAssign(tmp, a))
+                arg_vars.append(tmp)
+            else:
+                arg_vars.append(a)
+
+        initial = self._print(convert_to_literal(0, expr.dtype))
+        tmp_additional_code = self._additional_code
+        self._additional_code = ''
+        lhs = self._print(lhs_var)
+
+        loop_scope = self.scope.create_new_loop_scope()
+        iter_var_name = loop_scope.get_new_name()
+        iter_var = Variable(PythonNativeInt(), iter_var_name)
+
+        x1, x2 = arg_vars
+        if with_conjugate and x1.dtype.primitive_type is PrimitiveComplexType():
+            node = self._print(AugAssign(lhs_var, '+', PyccelMul(PythonConjugate(x1[iter_var]), x2[iter_var])))
+        else:
+            node = self._print(AugAssign(lhs_var, '+', PyccelMul(x1[iter_var], x2[iter_var])))
+        body = self._additional_code + node
+        self._additional_code = tmp_additional_code
+
+        code = prefix + (f'{lhs} = {initial};\n'
+                f'for (c_range({iter_var_name}, {self._print(x1.shape[axis])})) {{\n'
+                f'{body}'
+                 '}\n')
+        if getattr(lhs_var, 'is_temp', False):
+            self._additional_code += code
+            return lhs
+        else:
+            return code
 
     # ============ Elements ============ #
 
@@ -1886,11 +1961,12 @@ class CCodePrinter(CodePrinter):
         arg = expr.arg
         if isinstance(arg.class_type, (NumpyNDArrayType, HomogeneousTupleType)):
             idx = self._print(expr.index)
+            cast_code = f'({self.get_c_type(PythonNativeInt())})'
             if self.is_c_pointer(arg):
                 arg_code = self._print(ObjectAddress(arg))
-                return f'{arg_code}->shape[{idx}]'
+                return f'{cast_code}{arg_code}->shape[{idx}]'
             arg_code = self._print(arg)
-            return f'{arg_code}.shape[{idx}]'
+            return f'{cast_code}{arg_code}.shape[{idx}]'
         elif isinstance(arg.class_type, (HomogeneousListType, HomogeneousSetType, DictType)):
             c_type = self.get_c_type(arg.class_type)
             arg_code = self._print(ObjectAddress(arg))
@@ -2388,6 +2464,39 @@ class CCodePrinter(CodePrinter):
                 f'{c_1} = {a_2} * {b_0} - {a_0} * {b_2};\n'
                 f'{c_2} = {a_0} * {b_1} - {a_1} * {b_0};\n')
 
+    def _print_NumpyVecdot(self, expr):
+        return self._inline_vecdot(expr, expr.x1, expr.x2, expr.axis[0], with_conjugate = True)
+
+    def _print_NumpyMatmul(self, expr):
+        if expr.rank == 0:
+            return self._inline_vecdot(expr, expr.a, expr.b, 0, with_conjugate = False)
+        self.add_import(c_imports['pyc_math_c'])
+        a_code = self._print(expr.a)
+        b_code = self._print(expr.b)
+        assign = expr.get_user_nodes(Assign)
+        if assign:
+            out_code = self._print(assign[0].lhs)
+        else:
+            errors.report('The result of a matrix multiplication must be saved into a variable',
+                          symbol=expr, severity='error')
+            return ''
+        dtype = expr.dtype
+        if dtype.primitive_type is PrimitiveBooleanType():
+            errors.report('Boolean matrix multiplication is not implemented',
+                          symbol=expr, severity='error')
+        elif len({expr.a.dtype, expr.b.dtype, expr.dtype}) != 1:
+            errors.report('Matrix multiplication is only implemented in C when all arguments and results have the same type',
+                          symbol=expr, severity='error')
+        if expr.b.rank == 1:
+            dtype = self.get_c_type(dtype)
+            return f'pyc_matvecmul_{dtype}({out_code}, {a_code}, {b_code});\n'
+        elif expr.a.rank == 1:
+            dtype = self.get_c_type(dtype)
+            return f'pyc_vecmatmul_{dtype}({out_code}, {a_code}, {b_code});\n'
+        else:
+            array_type = self.get_c_type(expr.class_type)
+            return f'pyc_matmul_{array_type}({out_code}, {a_code}, {b_code});\n'
+
     def _print_Interface(self, expr):
         return ''.join(self._print(f) for f in expr.functions)
 
@@ -2661,7 +2770,7 @@ class CCodePrinter(CodePrinter):
         # Inhomogeneous tuples are unravelled and therefore do not exist in the c printer
         if isinstance(rhs, (NumpyArray, PythonTuple)):
             return self.copy_NumpyArray_Data(lhs, rhs)
-        if isinstance(rhs, NumpyReduction) and rhs.arg.rank:
+        if isinstance(rhs, NumpyMatmul) or (isinstance(rhs, NumpyReduction) and rhs.arg.rank):
             return self._print(rhs)
         if isinstance(rhs, (NumpyFull)):
             return self.arrayFill(expr)